requirements.txt
alpha-vantage==2.0.0
asn1crypto==0.24.0
attrs==18.1.0
autobahn==18.5.2
Automat==0.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
Django==2.0.5
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
newsapi-python==0.2.2
numpy==1.14.4
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.13.0
Robinhood==1.0.1
service-identity==17.0.0
simplejson==3.15.0
six==1.11.0
TA-Lib==0.4.17
Twisted==18.4.0
txaio==2.10.0
unicode==2.6
urllib3==1.22
watson-developer-cloud==1.3.5
zope.interface==4.5.0
